Update Readme to include migration option
After installing this gem and running the commands outlined in the readme I realised that migration files weren't being generated (gem version `0.6.0`). It appears that the migration argument is required for the migration files to be generated. Is this expected? Based on the readme I would have expected migrations to be generated by default. 

If you'd prefer that I update the default to generate migrations when the migration option isn't specified I can close this PR and open a separate one.